refactor(ui): address eighth self-review round
- task-grouped-view: clear selectedIds when hideCompleted toggles to prevent
  stale selection causing silently-wrong bulk cancel counts
- process-card: import formatStepName from @/lib/format (canonical source)
- agent-run-detail: remove local formatStepName, import from @/lib/format
- table-editor-view: remove re-export of deriveInitials (no longer needed)
- table-editor-avatar.test: import deriveInitials directly from @/lib/format
