+++ packages/platform-ui/src/components/agents/agent-run-detail.tsx
@@ -10,6 +10,7 @@ import type { AgentRun, ProcessInstance } from '@mediforce/platform-core';
 import { ConfidenceBadge } from './confidence-badge';
 import { AutonomyBadge } from './autonomy-badge';
 import { cn } from '@/lib/utils';
+import { formatStepName } from '@/lib/format';
 
 const STATUS_STYLES: Record<string, string> = {
   completed: 'bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-300',
@@ -71,13 +72,6 @@ function formatDuration(start: string, end: string | null): string {
   return `${(ms / 1000).toFixed(1)}s`;
 }
 
-/** Convert "supply-review" to "Supply Review" */
-function formatStepName(stepId: string): string {
-  return stepId
-    .split('-')
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(' ');
-}
 
 /** Render a record as flat key-value table rows (primitives) + collapsible JSON (objects/arrays) */
 function LightFormattedData({ data }: { data: Record<string, unknown> }) {
